Added source normalizer, turning tabs into spaces.
The normalizer will work on file patterns, affecting any files that
match the given pattern in the directory where it is executed. The usage
is:

    normalize file-pattern

Where "file-pattern" is the name of the file to normalize, or a file
pattern such as *.c or *.py etc.

Leading tabs will be replaced by four spaces.